Rename files to have pascal cased names
